requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.6
dj-database-url==0.5.0
Django==3.2.7
django-ipware==4.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
idna==3.2
lxml==4.6.3
passlib==1.7.4
Pillow==8.3.2
psycopg2==2.9.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
python-decouple==3.5
pytz==2021.1
requests==2.26.0
soupsieve==2.2.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uWSGI==2.0.20
validate-email==1.3
whitenoise==5.3.0
